requirements.txt
earthengine-api
geemap
numpy
Pillow
opencv-python
scipy
scikit-learn
